Boostrap niche overlap indices

Bootstrap niche overlap indices

niche.overlap.boot(mat, method = c("pianka", "schoener", "petraitis",
    "czech", "morisita", "levins"), times = 1000, quant = c(0.025, 0.975))

Arguments

mat
standard community matrix.

method
character string specifying the index.

times
Interger, representing the number of bootstrap samples to generate.

quant
Quantile of the bootstrap values.

Details

This function bootstraps the following niche overlap indices between each pair of species: \ schoener: Schoener's niche overlap index\ petraitis: Petraitis' niche overlap index\ czech: Czechanowski index \ morisita: Morisita's overlap index\ levins: Levin's overlap index\ see more information from Gotelli, N(2009).\

Value

a data frame with each row for each pair of species the columns are "Observed", "Boot mean", "Boot std", "Boot CI1", "Boot CI2", "times"

References

Zhang Jin-tun,(2004 ) Quantitative Ecology, Science Press, Beijing\

Nicholas J. Gotelli. 2000. Null model analysis of species co-occurrence patterns. Ecology 81:2606-2621. http://esapubs.org/archive/ecol/E081/022/EcoSim

Author

Jinlong Zhang jinlongzhang01@gmail.com

See also

Examples

data(datasample)
niche.overlap.boot(datasample[,1:4], method = "pianka")

#>                                            Observed Boot mean Boot std Boot CI1
#> Castanopsis.eyrei-Schima.superba              0.887     0.894    0.044    0.794
#> Castanopsis.eyrei-Rhododendron.ovatum         0.820     0.831    0.088    0.633
#> Castanopsis.eyrei-Cyclobalanopsis.glauca      0.551       NaN       NA    0.273
#> Schima.superba-Rhododendron.ovatum            0.693     0.725    0.146    0.423
#> Schima.superba-Cyclobalanopsis.glauca         0.491       NaN       NA    0.295
#> Rhododendron.ovatum-Cyclobalanopsis.glauca    0.253       NaN       NA    0.090
#>                                            Boot CI2 times
#> Castanopsis.eyrei-Schima.superba              0.972  1000
#> Castanopsis.eyrei-Rhododendron.ovatum         0.972  1000
#> Castanopsis.eyrei-Cyclobalanopsis.glauca      0.855  1000
#> Schima.superba-Rhododendron.ovatum            0.940  1000
#> Schima.superba-Cyclobalanopsis.glauca         0.801  1000
#> Rhododendron.ovatum-Cyclobalanopsis.glauca    0.577  1000

niche.overlap.boot(datasample[,1:4], method = "schoener")

#>                                            Observed Boot mean Boot std Boot CI1
#> Castanopsis.eyrei-Schima.superba              0.763     0.785    0.060    0.657
#> Castanopsis.eyrei-Rhododendron.ovatum         0.725     0.732    0.093    0.515
#> Castanopsis.eyrei-Cyclobalanopsis.glauca      0.416     0.446    0.123    0.210
#> Schima.superba-Rhododendron.ovatum            0.645     0.659    0.123    0.399
#> Schima.superba-Cyclobalanopsis.glauca         0.388       NaN       NA    0.157
#> Rhododendron.ovatum-Cyclobalanopsis.glauca    0.247       NaN       NA    0.099
#>                                            Boot CI2 times
#> Castanopsis.eyrei-Schima.superba              0.900  1000
#> Castanopsis.eyrei-Rhododendron.ovatum         0.887  1000
#> Castanopsis.eyrei-Cyclobalanopsis.glauca      0.715  1000
#> Schima.superba-Rhododendron.ovatum            0.858  1000
#> Schima.superba-Cyclobalanopsis.glauca         0.750  1000
#> Rhododendron.ovatum-Cyclobalanopsis.glauca    0.568  1000

niche.overlap.boot(datasample[,1:4], method = "czech")

#>                                            Observed Boot mean Boot std Boot CI1
#> Castanopsis.eyrei-Schima.superba              0.763     0.786    0.060    0.662
#> Castanopsis.eyrei-Rhododendron.ovatum         0.725     0.730    0.087    0.525
#> Castanopsis.eyrei-Cyclobalanopsis.glauca      0.416     0.442    0.130    0.185
#> Schima.superba-Rhododendron.ovatum            0.645     0.650    0.127    0.370
#> Schima.superba-Cyclobalanopsis.glauca         0.388     0.427    0.138    0.174
#> Rhododendron.ovatum-Cyclobalanopsis.glauca    0.247       NaN       NA    0.095
#>                                            Boot CI2 times
#> Castanopsis.eyrei-Schima.superba              0.905  1000
#> Castanopsis.eyrei-Rhododendron.ovatum         0.880  1000
#> Castanopsis.eyrei-Cyclobalanopsis.glauca      0.706  1000
#> Schima.superba-Rhododendron.ovatum            0.857  1000
#> Schima.superba-Cyclobalanopsis.glauca         0.740  1000
#> Rhododendron.ovatum-Cyclobalanopsis.glauca    0.572  1000

niche.overlap.boot(datasample[,1:4], method = "levins")

#>                                            Observed Boot mean Boot std Boot CI1
#> Castanopsis.eyrei-Schima.superba              0.921     0.929    0.051    0.823
#> Castanopsis.eyrei-Rhododendron.ovatum         0.947     0.952    0.045    0.864
#> Castanopsis.eyrei-Cyclobalanopsis.glauca      0.781     0.790    0.163    0.436
#> Schima.superba-Rhododendron.ovatum            0.772     0.809    0.196    0.432
#> Schima.superba-Cyclobalanopsis.glauca         0.671       NaN       NA    0.510
#> Rhododendron.ovatum-Cyclobalanopsis.glauca    0.311     0.383    0.209    0.094
#>                                            Boot CI2 times
#> Castanopsis.eyrei-Schima.superba              1.022  1000
#> Castanopsis.eyrei-Rhododendron.ovatum         1.050  1000
#> Castanopsis.eyrei-Cyclobalanopsis.glauca      1.071  1000
#> Schima.superba-Rhododendron.ovatum            1.118  1000
#> Schima.superba-Cyclobalanopsis.glauca         0.926  1000
#> Rhododendron.ovatum-Cyclobalanopsis.glauca    1.035  1000