Species association between each pair of species

Compute species association between each pair of species.

sp.pair(matr)

Arguments

matr
Standard community matrix, with rows representing sites and columns representing species.

Details

If a, b, c, d denote the co-occurrence the two species A and B, where:

a = number of plots occupied both by A and B.

b = number of plots only have A.

c = number of plots only have B.

d = number of plots without A or B.

N = a+b+c+d

Then, it is possible to compute:

Chi square (Yate's correction): chi^{2}=((((a*d-b*c)-0.5*N)^2)*N)/(a+b)*(a+c)*(b+d)*(c+d)

V ratio: V = ((a+d)-(b+c))/(a + b + c + d)

Jaccard index: Jaccard =a/(a + b + c)

Ochiai index: Ochiai = a/sqrt((a+b)*(a+c))

Dice index: Dice = 2*a/(2*a + b + c)

The Association Coefficient(AC):

if a*d >= b*c:

AC = (a*d - b*c)/((a+b)*(b+d))

if a*d < b*c and a <= d:

AC = (a*d - b*c)/((a+b)*(a+c))

if a*d < b*c and a > d:

AC = (a*d - b*c)/((b+d)*(c+d))

Point correlation coefficient

(PCC):

PCC = {a*d-b*c}/{(a+b)*(a+c)*(c+d)*(b+d)}

Value

chisq

Chi Square matrix

V

V positive or negative association

Ochiai

Ochiai's index

Dice

Dice's index

Jaccard

Jaccard's index

Pearson

Pearson's correlation coefficient

Spearman

Spearman's rank correlation coefficient

PCC

Point correlation coefficient

AC

Association coefficient

References

HURLBERT, S. H. (1969). A coefficient of interspecific association. Ecology, 50(1), 1-9.

WANG, B. S., & PENG S. L. (1985). Studies on the Measuring Techniques of Interspecific Association of Lower-Subtropical Evergreen-Broadleaved Forests. I. The Exploration and the Revision on the Measuring Formulas of Interspecific Association. Chinese Journal of Plant Ecology, 9(4), 274-285.

JIAN, M. F., LIU, Q. J., ZHU, D., & YOU, H. (2009). Inter-specific correlations among dominant populations of tree layer species in evergreen broad-leaved forest in Jiulianshan Mountain of subtropical China. Chinese Journal of Plant Ecology, 33(4), 672-680.

ZHOU, X. Y., WANG, B. S., LI, M. G., & ZAN, Q. J. (2000). An analysis of interspecific associations in secondary succession forest communities in Heishiding Natural Reserve, Guangdong Province. Chinese Journal of Plant Ecology, 24(3), 332-339

Author

Jinlong Zhang jinlongzhang01@gmail.com

See also

Examples

data(testdata)
spmatrix <- data2mat(testdata)
result <- sp.pair(spmatrix)